rainbow: re-organized rainbow files
